feat: exchangeAmount를 Double 대신 Integer로 바꾸니 정상적으로 응답하여 이렇게 변경

+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/dto/finance/exchange/ExchangeRequestDto.java
@@ -19,5 +19,5 @@ public class ExchangeRequestDto extends FinanceApiRequestDto {
 	/**
 	 * Double 이지만 단위는 10이다.
 	 */
-	Double exchangeAmount;
+	Integer exchangeAmount;
 }
